tests/wpt/mozilla/tests/css/css-property-listing.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>Web-exposed CSS properties</title>
<link rel="author" title="Oriol Brufau" href="mailto:obrufau@igalia.com">
<meta assert="This test documents Servo's web-exposed CSS properties.">
<!-- Inspired by Blink's third_party/blink/web_tests/webexposed/css-property-listing.html -->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
let expected_longhands = [
  "align-content",
  "align-items",
  "align-self",
  "animation-delay",
  "animation-direction",
  "animation-duration",
  "animation-fill-mode",
  "animation-iteration-count",
  "animation-name",
  "animation-play-state",
  "animation-timing-function",
  "aspect-ratio",
  "backface-visibility",
  "background-attachment",
  "background-clip",
  "background-color",
  "background-image",
  "background-origin",
  "background-position-x",
  "background-position-y",
  "background-repeat",
  "background-size",
  "baseline-source",
  "block-size",
  "border-block-end-color",
  "border-block-end-style",
  "border-block-end-width",
  "border-block-start-color",
  "border-block-start-style",
  "border-block-start-width",
  "border-bottom-color",
  "border-bottom-left-radius",
  "border-bottom-right-radius",
  "border-bottom-style",
  "border-bottom-width",
  "border-collapse",
  "border-end-end-radius",
  "border-end-start-radius",
  "border-image-outset",
  "border-image-repeat",
  "border-image-slice",
  "border-image-source",
  "border-image-width",
  "border-inline-end-color",
  "border-inline-end-style",
  "border-inline-end-width",
  "border-inline-start-color",
  "border-inline-start-style",
  "border-inline-start-width",
  "border-left-color",
  "border-left-style",
  "border-left-width",
  "border-right-color",
  "border-right-style",
  "border-right-width",
  "border-spacing",
  "border-start-end-radius",
  "border-start-start-radius",
  "border-top-color",
  "border-top-left-radius",
  "border-top-right-radius",
  "border-top-style",
  "border-top-width",
  "bottom",
  "box-shadow",
  "box-sizing",
  "caption-side",
  "caret-color",
  "clear",
  "clip",
  "clip-path",
  "color",
  "column-count",
  "column-gap",
  "column-span",
  "column-width",
  "container-name",
  "container-type",
  "content",
  "cursor",
  "direction",
  "display",
  "empty-cells",
  "filter",
  "flex-basis",
  "flex-direction",
  "flex-grow",
  "flex-shrink",
  "flex-wrap",
  "float",
  "font-family",
  "font-language-override",
  "font-optical-sizing",
  "font-size",
  "font-stretch",
  "font-style",
  "font-synthesis-weight",
  "font-variant-caps",
  "font-variation-settings",
  "font-weight",
  "grid-auto-columns",
  "grid-auto-flow",
  "grid-auto-rows",
  "grid-column-end",
  "grid-column-start",
  "grid-row-end",
  "grid-row-start",
  "grid-template-areas",
  "grid-template-columns",
  "grid-template-rows",
  "height",
  "image-rendering",
  "inline-size",
  "inset-block-end",
  "inset-block-start",
  "inset-inline-end",
  "inset-inline-start",
  "isolation",
  "justify-content",
  "justify-items",
  "justify-self",
  "left",
  "letter-spacing",
  "line-break",
  "line-height",
  "list-style-image",
  "list-style-position",
  "list-style-type",
  "margin-block-end",
  "margin-block-start",
  "margin-bottom",
  "margin-inline-end",
  "margin-inline-start",
  "margin-left",
  "margin-right",
  "margin-top",
  "max-block-size",
  "max-height",
  "max-inline-size",
  "max-width",
  "min-block-size",
  "min-height",
  "min-inline-size",
  "min-width",
  "mix-blend-mode",
  "object-fit",
  "object-position",
  "opacity",
  "order",
  "outline-color",
  "outline-offset",
  "outline-style",
  "outline-width",
  "overflow-block",
  "overflow-clip-margin",
  "overflow-inline",
  "overflow-wrap",
  "overflow-x",
  "overflow-y",
  "padding-block-end",
  "padding-block-start",
  "padding-bottom",
  "padding-inline-end",
  "padding-inline-start",
  "padding-left",
  "padding-right",
  "padding-top",
  "perspective",
  "perspective-origin",
  "pointer-events",
  "position",
  "quotes",
  "right",
  "rotate",
  "row-gap",
  "scale",
  "table-layout",
  "text-align",
  "text-align-last",
  "text-decoration-color",
  "text-decoration-line",
  "text-decoration-style",
  "text-indent",
  "text-justify",
  "text-rendering",
  "text-shadow",
  "text-transform",
  "text-wrap-mode",
  "top",
  "transform",
  "transform-origin",
  "transform-style",
  "transition-behavior",
  "transition-delay",
  "transition-duration",
  "transition-property",
  "transition-timing-function",
  "translate",
  "unicode-bidi",
  "vertical-align",
  "visibility",
  "white-space-collapse",
  "width",
  "will-change",
  "word-break",
  "word-spacing",
  "z-index",
  "-webkit-text-security",
];

let expected_shorthands = {
  "all": [
    "border-block-start-color",
    "border-block-start-style",
    "border-block-start-width",
    "border-block-end-color",
    "border-block-end-style",
    "border-block-end-width",
    "border-inline-start-color",
    "border-inline-start-style",
    "border-inline-start-width",
    "border-inline-end-color",
    "border-inline-end-style",
    "border-inline-end-width",
    "border-start-start-radius",
    "border-start-end-radius",
    "border-end-start-radius",
    "border-end-end-radius",
    "overflow-block",
    "overflow-inline",
    "margin-block-start",
    "margin-block-end",
    "margin-inline-start",
    "margin-inline-end",
    "padding-block-start",
    "padding-block-end",
    "padding-inline-start",
    "padding-inline-end",
    "inset-block-start",
    "inset-block-end",
    "inset-inline-start",
    "inset-inline-end",
    "block-size",
    "min-block-size",
    "max-block-size",
    "inline-size",
    "min-inline-size",
    "max-inline-size",
    "background-color",
    "background-image",
    "background-position-x",
    "background-position-y",
    "background-repeat",
    "background-attachment",
    "background-clip",
    "background-origin",
    "background-size",
    "border-top-color",
    "border-top-style",
    "border-top-width",
    "border-right-color",
    "border-right-style",
    "border-right-width",
    "border-bottom-color",
    "border-bottom-style",
    "border-bottom-width",
    "border-left-color",
    "border-left-style",
    "border-left-width",
    "border-top-left-radius",
    "border-top-right-radius",
    "border-bottom-right-radius",
    "border-bottom-left-radius",
    "border-image-source",
    "border-image-outset",
    "border-image-repeat",
    "border-image-width",
    "border-image-slice",
    "display",
    "position",
    "float",
    "clear",
    "vertical-align",
    "baseline-source",
    "overflow-y",
    "overflow-x",
    "transform",
    "rotate",
    "scale",
    "translate",
    "isolation",
    "perspective",
    "perspective-origin",
    "backface-visibility",
    "transform-style",
    "transform-origin",
    "container-type",
    "container-name",
    "will-change",
    "column-width",
    "column-count",
    "column-span",
    "content",
    "opacity",
    "box-shadow",
    "clip",
    "filter",
    "mix-blend-mode",
    "font-family",
    "font-style",
    "font-variant-caps",
    "font-weight",
    "font-size",
    "font-synthesis-weight",
    "font-stretch",
    "font-variation-settings",
    "font-language-override",
    "font-optical-sizing",
    "line-height",
    "visibility",
    "image-rendering",
    "border-collapse",
    "empty-cells",
    "caption-side",
    "border-spacing",
    "color",
    "text-transform",
    "text-indent",
    "overflow-wrap",
    "word-break",
    "text-justify",
    "text-align-last",
    "text-align",
    "letter-spacing",
    "word-spacing",
    "white-space-collapse",
    "text-shadow",
    "line-break",
    "text-rendering",
    "-webkit-text-security",
    "text-wrap-mode",
    "cursor",
    "pointer-events",
    "caret-color",
    "list-style-position",
    "list-style-type",
    "list-style-image",
    "quotes",
    "margin-top",
    "margin-right",
    "margin-bottom",
    "margin-left",
    "overflow-clip-margin",
    "outline-color",
    "outline-style",
    "outline-width",
    "outline-offset",
    "padding-top",
    "padding-right",
    "padding-bottom",
    "padding-left",
    "top",
    "right",
    "bottom",
    "left",
    "z-index",
    "flex-direction",
    "flex-wrap",
    "justify-content",
    "align-content",
    "align-items",
    "justify-items",
    "flex-grow",
    "flex-shrink",
    "align-self",
    "justify-self",
    "order",
    "flex-basis",
    "height",
    "min-height",
    "max-height",
    "width",
    "min-width",
    "max-width",
    "box-sizing",
    "object-fit",
    "object-position",
    "grid-row-start",
    "grid-row-end",
    "grid-auto-rows",
    "grid-template-rows",
    "grid-column-start",
    "grid-column-end",
    "grid-auto-columns",
    "grid-template-columns",
    "grid-auto-flow",
    "grid-template-areas",
    "column-gap",
    "row-gap",
    "aspect-ratio",
    "clip-path",
    "table-layout",
    "text-decoration-line",
    "text-decoration-style",
    "text-decoration-color",
    "transition-duration",
    "transition-timing-function",
    "transition-property",
    "transition-delay",
    "transition-behavior",
    "animation-name",
    "animation-duration",
    "animation-timing-function",
    "animation-iteration-count",
    "animation-direction",
    "animation-play-state",
    "animation-fill-mode",
    "animation-delay",
  ],
  "animation": [
    "animation-name",
    "animation-duration",
    "animation-timing-function",
    "animation-delay",
    "animation-iteration-count",
    "animation-direction",
    "animation-fill-mode",
    "animation-play-state",
  ],
  "background": [
    "background-color",
    "background-position-x",
    "background-position-y",
    "background-repeat",
    "background-attachment",
    "background-image",
    "background-size",
    "background-origin",
    "background-clip",
  ],
  "background-position": [
    "background-position-x",
    "background-position-y",
  ],
  "border": [
    "border-top-width",
    "border-top-style",
    "border-top-color",
    "border-right-width",
    "border-right-style",
    "border-right-color",
    "border-bottom-width",
    "border-bottom-style",
    "border-bottom-color",
    "border-left-width",
    "border-left-style",
    "border-left-color",
    "border-image-outset",
    "border-image-repeat",
    "border-image-slice",
    "border-image-source",
    "border-image-width",
  ],
  "border-block": [
    "border-block-start-width",
    "border-block-end-width",
    "border-block-start-style",
    "border-block-end-style",
    "border-block-start-color",
    "border-block-end-color",
  ],
  "border-block-color": [
    "border-block-start-color",
    "border-block-end-color",
  ],
  "border-block-end": [
    "border-block-end-width",
    "border-block-end-style",
    "border-block-end-color",
  ],
  "border-block-start": [
    "border-block-start-width",
    "border-block-start-style",
    "border-block-start-color",
  ],
  "border-block-style": [
    "border-block-start-style",
    "border-block-end-style",
  ],
  "border-block-width": [
    "border-block-start-width",
    "border-block-end-width",
  ],
  "border-bottom": [
    "border-bottom-width",
    "border-bottom-style",
    "border-bottom-color",
  ],
  "border-color": [
    "border-top-color",
    "border-right-color",
    "border-bottom-color",
    "border-left-color",
  ],
  "border-image": [
    "border-image-outset",
    "border-image-repeat",
    "border-image-slice",
    "border-image-source",
    "border-image-width",
  ],
  "border-inline": [
    "border-inline-start-width",
    "border-inline-end-width",
    "border-inline-start-style",
    "border-inline-end-style",
    "border-inline-start-color",
    "border-inline-end-color",
  ],
  "border-inline-color": [
    "border-inline-start-color",
    "border-inline-end-color",
  ],
  "border-inline-end": [
    "border-inline-end-width",
    "border-inline-end-style",
    "border-inline-end-color",
  ],
  "border-inline-start": [
    "border-inline-start-width",
    "border-inline-start-style",
    "border-inline-start-color",
  ],
  "border-inline-style": [
    "border-inline-start-style",
    "border-inline-end-style",
  ],
  "border-inline-width": [
    "border-inline-start-width",
    "border-inline-end-width",
  ],
  "border-left": [
    "border-left-width",
    "border-left-style",
    "border-left-color",
  ],
  "border-radius": [
    "border-top-left-radius",
    "border-top-right-radius",
    "border-bottom-right-radius",
    "border-bottom-left-radius",
  ],
  "border-right": [
    "border-right-width",
    "border-right-style",
    "border-right-color",
  ],
  "border-style": [
    "border-top-style",
    "border-right-style",
    "border-bottom-style",
    "border-left-style",
  ],
  "border-top": [
    "border-top-width",
    "border-top-style",
    "border-top-color",
  ],
  "border-width": [
    "border-top-width",
    "border-right-width",
    "border-bottom-width",
    "border-left-width",
  ],
  "columns": [
    "column-width",
    "column-count",
  ],
  "flex": [
    "flex-grow",
    "flex-shrink",
    "flex-basis",
  ],
  "flex-flow": [
    "flex-direction",
    "flex-wrap",
  ],
  "font": [
    "font-style",
    "font-variant-caps",
    "font-weight",
    "font-stretch",
    "font-size",
    "line-height",
    "font-family",
    "font-optical-sizing",
    "font-variation-settings",
  ],
  "font-variant": [
    "font-variant-caps",
  ],
  "gap": [
    "row-gap",
    "column-gap",
  ],
  "grid": [
    "grid-template-rows",
    "grid-template-columns",
    "grid-template-areas",
    "grid-auto-rows",
    "grid-auto-columns",
    "grid-auto-flow",
  ],
  "grid-area": [
    "grid-row-start",
    "grid-row-end",
    "grid-column-start",
    "grid-column-end",
  ],
  "grid-column": [
    "grid-column-start",
    "grid-column-end",
  ],
  "grid-gap": [
    "row-gap",
    "column-gap",
  ],
  "grid-row": [
    "grid-row-start",
    "grid-row-end",
  ],
  "grid-template": [
    "grid-template-rows",
    "grid-template-columns",
    "grid-template-areas",
  ],
  "inset": [
    "top",
    "right",
    "bottom",
    "left",
  ],
  "inset-block": [
    "inset-block-start",
    "inset-block-end",
  ],
  "inset-inline": [
    "inset-inline-start",
    "inset-inline-end",
  ],
  "list-style": [
    "list-style-position",
    "list-style-image",
    "list-style-type",
  ],
  "margin": [
    "margin-top",
    "margin-right",
    "margin-bottom",
    "margin-left",
  ],
  "margin-block": [
    "margin-block-start",
    "margin-block-end",
  ],
  "margin-inline": [
    "margin-inline-start",
    "margin-inline-end",
  ],
  "outline": [
    "outline-color",
    "outline-style",
    "outline-width",
  ],
  "overflow": [
    "overflow-x",
    "overflow-y",
  ],
  "padding": [
    "padding-top",
    "padding-right",
    "padding-bottom",
    "padding-left",
  ],
  "padding-block": [
    "padding-block-start",
    "padding-block-end",
  ],
  "padding-inline": [
    "padding-inline-start",
    "padding-inline-end",
  ],
  "place-content": [
    "align-content",
    "justify-content",
  ],
  "place-items": [
    "align-items",
    "justify-items",
  ],
  "place-self": [
    "align-self",
    "justify-self",
  ],
  "text-decoration": [
    "text-decoration-color",
    "text-decoration-line",
    "text-decoration-style",
  ],
  "transition": [
    "transition-property",
    "transition-duration",
    "transition-timing-function",
    "transition-delay",
    "transition-behavior",
  ],
  "white-space": [
    "text-wrap-mode",
    "white-space-collapse",
  ],
  "-moz-animation": [
    "animation-name",
    "animation-duration",
    "animation-timing-function",
    "animation-delay",
    "animation-iteration-count",
    "animation-direction",
    "animation-fill-mode",
    "animation-play-state",
  ],
  "-moz-border-image": [
    "border-image-outset",
    "border-image-repeat",
    "border-image-slice",
    "border-image-source",
    "border-image-width",
  ],
  "-moz-transition": [
    "transition-property",
    "transition-duration",
    "transition-timing-function",
    "transition-delay",
    "transition-behavior",
  ],
  "-webkit-animation": [
    "animation-name",
    "animation-duration",
    "animation-timing-function",
    "animation-delay",
    "animation-iteration-count",
    "animation-direction",
    "animation-fill-mode",
    "animation-play-state",
  ],
  "-webkit-border-image": [
    "border-image-outset",
    "border-image-repeat",
    "border-image-slice",
    "border-image-source",
    "border-image-width",
  ],
  "-webkit-border-radius": [
    "border-top-left-radius",
    "border-top-right-radius",
    "border-bottom-right-radius",
    "border-bottom-left-radius",
  ],
  "-webkit-flex": [
    "flex-grow",
    "flex-shrink",
    "flex-basis",
  ],
  "-webkit-flex-flow": [
    "flex-direction",
    "flex-wrap",
  ],
  "-webkit-perspective": [
    "perspective",
  ],
  "-webkit-transform": [
    "transform",
  ],
  "-webkit-transition": [
    "transition-property",
    "transition-duration",
    "transition-timing-function",
    "transition-delay",
    "transition-behavior",
  ],
};

let expected_aliases = {
  "grid-column-gap": "column-gap",
  "grid-row-gap": "row-gap",
  "word-wrap": "overflow-wrap",
  "-moz-animation-delay": "animation-delay",
  "-moz-animation-direction": "animation-direction",
  "-moz-animation-duration": "animation-duration",
  "-moz-animation-fill-mode": "animation-fill-mode",
  "-moz-animation-iteration-count": "animation-iteration-count",
  "-moz-animation-name": "animation-name",
  "-moz-animation-play-state": "animation-play-state",
  "-moz-animation-timing-function": "animation-timing-function",
  "-moz-backface-visibility": "backface-visibility",
  "-moz-box-sizing": "box-sizing",
  "-moz-font-language-override": "font-language-override",
  "-moz-perspective": "perspective",
  "-moz-perspective-origin": "perspective-origin",
  "-moz-transform": "transform",
  "-moz-transform-origin": "transform-origin",
  "-moz-transform-style": "transform-style",
  "-moz-transition-delay": "transition-delay",
  "-moz-transition-duration": "transition-duration",
  "-moz-transition-property": "transition-property",
  "-moz-transition-timing-function": "transition-timing-function",
  "-webkit-align-content": "align-content",
  "-webkit-align-items": "align-items",
  "-webkit-align-self": "align-self",
  "-webkit-animation-delay": "animation-delay",
  "-webkit-animation-direction": "animation-direction",
  "-webkit-animation-duration": "animation-duration",
  "-webkit-animation-fill-mode": "animation-fill-mode",
  "-webkit-animation-iteration-count": "animation-iteration-count",
  "-webkit-animation-name": "animation-name",
  "-webkit-animation-play-state": "animation-play-state",
  "-webkit-animation-timing-function": "animation-timing-function",
  "-webkit-backface-visibility": "backface-visibility",
  "-webkit-background-clip": "background-clip",
  "-webkit-background-origin": "background-origin",
  "-webkit-background-size": "background-size",
  "-webkit-border-bottom-left-radius": "border-bottom-left-radius",
  "-webkit-border-bottom-right-radius": "border-bottom-right-radius",
  "-webkit-border-top-left-radius": "border-top-left-radius",
  "-webkit-border-top-right-radius": "border-top-right-radius",
  "-webkit-box-shadow": "box-shadow",
  "-webkit-box-sizing": "box-sizing",
  "-webkit-clip-path": "clip-path",
  "-webkit-filter": "filter",
  "-webkit-flex-basis": "flex-basis",
  "-webkit-flex-direction": "flex-direction",
  "-webkit-flex-grow": "flex-grow",
  "-webkit-flex-shrink": "flex-shrink",
  "-webkit-flex-wrap": "flex-wrap",
  "-webkit-justify-content": "justify-content",
  "-webkit-order": "order",
  "-webkit-perspective-origin": "perspective-origin",
  "-webkit-transform-origin": "transform-origin",
  "-webkit-transform-style": "transform-style",
  "-webkit-transition-delay": "transition-delay",
  "-webkit-transition-duration": "transition-duration",
  "-webkit-transition-property": "transition-property",
  "-webkit-transition-timing-function": "transition-timing-function",
};

let { style } = document.createElement("foo");

let properties = [];
for (let property in style) {
  if (CSS.supports(property, "initial")) {
    properties.push(property);
  }
}
properties.sort((a,b) => {
  if (a === b) {
    return 0;
  }
  let a_is_prefixed = a.startsWith("-");
  let b_is_prefixed = b.startsWith("-");
  if (a_is_prefixed !== b_is_prefixed) {
    return b_is_prefixed ? -1 : 1;
  }
  return a < b ? -1 : 1;
});

let longhands = [];
let aliases = {};
let shorthands = {};
for (let property of properties) {
  style[property] = "var(--a)";
  if (style.length === 1) {
    let resulting_property = style[0];
    if (property === resulting_property) {
      longhands.push(property);
    } else if (style[property] === style[resulting_property]) {
      aliases[property] = resulting_property;
    } else {
      shorthands[property] = [resulting_property];
    }
  } else {
    shorthands[property] = [...style];
  }
  style.cssText = "";
}

test(() => {
  assert_array_equals(longhands, expected_longhands);
}, "Web-exposed CSS longhands");

test(() => {
  assert_object_equals(shorthands, expected_shorthands);
  assert_array_equals(Object.keys(shorthands), Object.keys(expected_shorthands));
}, "Web-exposed CSS shorthands");

test(() => {
  assert_object_equals(aliases, expected_aliases);
  assert_array_equals(Object.keys(aliases), Object.keys(expected_aliases));
}, "Web-exposed CSS aliases");
</script>
